"Fixed data-wow-duration attribute for naresh image and modified email link display in index.html"
Explanation: The code diff shows changes made to the index.html file. The first set of changes modifies the "home-img" section by fixing the repetition of the "data-wow-duration" attribute for the "naresh" image. The second set of changes modifies the display of the email link in the "mail" section by wrapping the email address in a separate div.